LayoutTests/imported/w3c/web-platform-tests/html/select/options-length-too-large-expected.txt

CONSOLE MESSAGE: Unable to expand the option list to length 4294967295 items. The maximum number of items allowed is 100000.
CONSOLE MESSAGE: Unable to expand the option list to length 100001 items. The maximum number of items allowed is 100000.
CONSOLE MESSAGE: Unable to expand the option list to length 4294967295 items. The maximum number of items allowed is 100000.


PASS select options.length too large
PASS select options.length too large 1
PASS select options.length too large 2
PASS select options.length too large 3
PASS select options.length too large 4
